+++ sphinx/application/data/concepts/repositories/concept-repository-actions/src/main/java/chat/sphinx/concept_repository_actions/ActionsRepository.kt
@@ -2,6 +2,7 @@ package chat.sphinx.concept_repository_actions
 
 import chat.sphinx.wrapper_action_track.action_wrappers.ContentConsumedHistoryItem
 import chat.sphinx.wrapper_common.feed.FeedId
+import kotlinx.coroutines.flow.MutableStateFlow
 
 interface ActionsRepository {
 
@@ -23,14 +24,14 @@ interface ActionsRepository {
         feedItemId: FeedId
     )
 
-    fun trackVideoConsumed(
+    fun trackMediaContentConsumed(
         feedItemId: FeedId,
-        history: java.util.ArrayList<ContentConsumedHistoryItem>
+        history: ArrayList<ContentConsumedHistoryItem>
     )
 
-    fun trackPodcastConsumed(
+    fun trackRecommendationsConsumed(
         feedItemId: FeedId,
-        history: java.util.ArrayList<ContentConsumedHistoryItem>
+        history: ArrayList<ContentConsumedHistoryItem>
     )
 
     fun trackMessageContent(
@@ -38,4 +39,7 @@ interface ActionsRepository {
     )
 
     fun syncActions()
+
+    val appLogsStateFlow: MutableStateFlow<String>
+    fun setAppLog(log: String)
 }