+++ sphinx/screens-detail/common-player/src/main/java/chat/sphinx/common_player/viewstate/EpisodePlayerViewState.kt
@@ -1,31 +1,36 @@
 package chat.sphinx.common_player.viewstate
 
 import chat.sphinx.concept_service_media.MediaPlayerServiceState
-import chat.sphinx.wrapper_feed.FeedRecommendation
 import chat.sphinx.wrapper_podcast.Podcast
 import chat.sphinx.wrapper_podcast.PodcastEpisode
 import io.matthewnelson.concept_views.viewstate.ViewState
 
-sealed class EpisodePlayerViewState: ViewState<EpisodePlayerViewState>() {
+sealed class RecommendationsPodcastPlayerViewState: ViewState<RecommendationsPodcastPlayerViewState>() {
 
-    object Idle: EpisodePlayerViewState()
-    object ServiceLoading: EpisodePlayerViewState()
-    object ServiceInactive: EpisodePlayerViewState()
+    object Idle: RecommendationsPodcastPlayerViewState()
+    object ServiceLoading: RecommendationsPodcastPlayerViewState()
+    object ServiceInactive: RecommendationsPodcastPlayerViewState()
 
-    class EpisodeLoaded(
-        val feedRecommendation: FeedRecommendation,
-    ): EpisodePlayerViewState()
+    sealed class PodcastViewState(
+        val podcast: Podcast,
+    ) : RecommendationsPodcastPlayerViewState() {
 
-    class LoadingEpisode(
-        val feedRecommendation: FeedRecommendation,
-    ): EpisodePlayerViewState()
+        class PodcastLoaded(
+            podcast: Podcast
+        ): PodcastViewState(podcast)
 
-    class EpisodePlayed(
-        val feedRecommendation: FeedRecommendation,
-    ): EpisodePlayerViewState()
+        class LoadingEpisode(
+            podcast: Podcast,
+            val episode: PodcastEpisode
+        ): PodcastViewState(podcast)
 
-    class MediaStateUpdate(
-        val feedRecommendation: FeedRecommendation,
-        val state: MediaPlayerServiceState.ServiceActive.MediaState
-    ): EpisodePlayerViewState()
+        class EpisodePlayed(
+            podcast: Podcast
+        ): PodcastViewState(podcast)
+
+        class MediaStateUpdate(
+            podcast: Podcast,
+            val state: MediaPlayerServiceState.ServiceActive.MediaState
+        ): PodcastViewState(podcast)
+    }
 }