+++ sphinx/screens/chats/chat-common/chat-common/src/main/java/chat/sphinx/chat_common/ui/ChatFragment.kt
@@ -10,6 +10,7 @@ import android.os.Bundle
 import android.os.ParcelFileDescriptor
 import android.text.Editable
 import android.text.TextWatcher
+import android.util.Log
 import android.view.*
 import android.view.inputmethod.InputMethodManager
 import android.widget.ImageView
@@ -133,6 +134,8 @@ abstract class ChatFragment<
 
     private val holderJobs: ArrayList<Job> = ArrayList(11)
     private val disposables: ArrayList<Disposable> = ArrayList(4)
+
+    private val timeTrackerStart = System.currentTimeMillis()
 
     override val chatFragmentContext: Context
         get() = binding.root.context
@@ -843,6 +846,8 @@ abstract class ChatFragment<
                 }
             })
         }
+        Log.d("TimeTracker", "Chat messages were displayed in ${System.currentTimeMillis() - timeTrackerStart} milliseconds")
+        viewModel.sendAppLog("- Chat messages were displayed in ${System.currentTimeMillis() - timeTrackerStart} milliseconds")
     }
 
     protected fun scrollToBottom(
@@ -1060,6 +1065,9 @@ abstract class ChatFragment<
 
                             textViewChatHeaderName.text = viewState.chatHeaderName
                             textViewChatHeaderLock.goneIfFalse(viewState.showLock)
+
+                            Log.d("TimeTracker", "Chat contact/tribe name was displayed in ${System.currentTimeMillis() - timeTrackerStart} milliseconds")
+                            viewModel.sendAppLog("- Chat contact/tribe name was displayed in ${System.currentTimeMillis() - timeTrackerStart} milliseconds")
 
                             imageViewChatHeaderMuted.apply {
                                 viewState.isMuted?.let { muted ->
@@ -1757,23 +1765,25 @@ abstract class ChatFragment<
     override fun onInteractionComplete() {
         viewModel.audioRecorderController.stopAudioRecording()
         viewModel.audioRecorderController.recordingTempFile?.let {
-            sendMessageBuilder.setAttachmentInfo(
-                AttachmentInfo(
-                    file = it,
-                    mediaType = MediaType.Audio(AudioRecorderController.AUDIO_FORMAT_MIME_TYPE),
-                    fileName = null,
-                    isLocalFile = true,
+            lifecycleScope.launch(viewModel.mainImmediate) {
+                sendMessageBuilder.setAttachmentInfo(
+                    AttachmentInfo(
+                        file = it,
+                        mediaType = MediaType.Audio(AudioRecorderController.AUDIO_FORMAT_MIME_TYPE),
+                        fileName = null,
+                        isLocalFile = true,
+                    )
                 )
-            )
 
-            viewModel.sendMessage(sendMessageBuilder)?.let {
-                // if it did not return null that means it was valid
-                viewModel.updateFooterViewState(FooterViewState.Default)
+                viewModel.sendMessage(sendMessageBuilder)?.let {
+                    // if it did not return null that means it was valid
+                    viewModel.updateFooterViewState(FooterViewState.Default)
 
-                sendMessageBuilder.clear()
-                viewModel.messageReplyViewStateContainer.updateViewState(MessageReplyViewState.ReplyingDismissed)
+                    sendMessageBuilder.clear()
+                    viewModel.messageReplyViewStateContainer.updateViewState(MessageReplyViewState.ReplyingDismissed)
+                }
+                viewModel.audioRecorderController.clear()
             }
-            viewModel.audioRecorderController.clear()
         }
     }
 }