+++ sphinx/screens/chats/chat-tribe/chat-tribe/src/main/java/chat/sphinx/chat_tribe/ui/TribeFeedViewModel.kt
@@ -158,13 +158,13 @@ internal class TribeFeedViewModel @Inject constructor(
             return
         }
 
-        @Exhaustive
         when (serviceState) {
             is MediaPlayerServiceState.ServiceActive.MediaState.Playing -> {
                 vs.podcast.playingEpisodeUpdate(
                     serviceState.episodeId,
                     serviceState.currentTime,
-                    serviceState.episodeDuration.toLong()
+                    serviceState.episodeDuration.toLong(),
+                    serviceState.speed
                 )
 
                 vs.adjustState(
@@ -230,6 +230,7 @@ internal class TribeFeedViewModel @Inject constructor(
                     podcastViewStateContainer.updateViewState(it)
                 }
             }
+            else -> {}
         }
     }
 
@@ -425,7 +426,7 @@ internal class TribeFeedViewModel @Inject constructor(
                 showLoading = true,
                 showPlayButton = true,
                 title = podcast.getCurrentEpisode().title.value,
-                subtitle = podcast.author?.value ?: podcast.title.value,
+                subtitle = podcast.author?.value ?: podcast.getCurrentEpisode().showTitle?.value ?: podcast.title.value,
                 imageUrl = podcast.imageToShow?.value,
                 playingProgress = 0,
                 clickPlayPause = clickPlayPause,
@@ -450,7 +451,7 @@ internal class TribeFeedViewModel @Inject constructor(
             showLoading = false,
             showPlayButton = !isPlaying,
             title = podcast.getCurrentEpisode().title.value,
-            subtitle = podcast.author?.value ?: podcast.title.value,
+            subtitle = podcast.author?.value ?: podcast.getCurrentEpisode().showTitle?.value ?: podcast.title.value,
             imageUrl = podcast.imageToShow?.value,
             playingProgress = playingProgress,
             clickPlayPause = clickPlayPause,