+++ sphinx/screens/dashboard/dashboard/src/main/java/chat/sphinx/dashboard/navigation/DashboardNavigator.kt
@@ -41,8 +41,9 @@ abstract class DashboardNavigator(
     abstract suspend fun toPodcastPlayerScreen(chatId: ChatId, feedId: FeedId, feedUrl: FeedUrl, currentEpisodeDuration: Long)
 
     abstract suspend fun toCommonPlayerScreen(
-        recommendations: List<String>,
-        recommendationId: String,
+        podcastId: FeedId,
+        episodeId: FeedId,
+        currentEpisodeDuration: Long
     )
 
 }