+++ sphinx/screens/dashboard/dashboard/src/main/java/chat/sphinx/dashboard/ui/feed/all/FeedAllViewModel.kt
@@ -1,8 +1,11 @@
 package chat.sphinx.dashboard.ui.feed.all
 
-import android.content.Context
+import androidx.fragment.app.FragmentActivity
 import androidx.lifecycle.viewModelScope
 import chat.sphinx.concept_repository_dashboard_android.RepositoryDashboardAndroid
+import chat.sphinx.concept_repository_feed.FeedRepository
+import chat.sphinx.concept_service_media.MediaPlayerServiceController
+import chat.sphinx.concept_service_media.UserAction
 import chat.sphinx.dashboard.navigation.DashboardNavigator
 import chat.sphinx.dashboard.ui.feed.FeedFollowingViewModel
 import chat.sphinx.dashboard.ui.feed.FeedRecommendationsViewModel
@@ -12,11 +15,12 @@ import chat.sphinx.wrapper_common.feed.FeedId
 import chat.sphinx.wrapper_common.feed.FeedType
 import chat.sphinx.wrapper_common.feed.FeedUrl
 import chat.sphinx.wrapper_feed.Feed
-import chat.sphinx.wrapper_feed.FeedRecommendation
-import chat.sphinx.wrapper_feed.toJson
+import chat.sphinx.wrapper_podcast.FeedRecommendation
+import chat.sphinx.wrapper_podcast.Podcast
 import com.squareup.moshi.Moshi
 import dagger.hilt.android.lifecycle.HiltViewModel
 import io.matthewnelson.android_feature_viewmodel.SideEffectViewModel
+import io.matthewnelson.android_feature_viewmodel.submitSideEffect
 import io.matthewnelson.android_feature_viewmodel.updateViewState
 import io.matthewnelson.concept_coroutines.CoroutineDispatchers
 import kotlinx.coroutines.flow.*
@@ -29,23 +33,51 @@ import javax.inject.Inject
 internal class FeedAllViewModel @Inject constructor(
     val dashboardNavigator: DashboardNavigator,
     private val repositoryDashboard: RepositoryDashboardAndroid<Any>,
+    private val feedRepository: FeedRepository,
+    private val mediaPlayerServiceController: MediaPlayerServiceController,
     val moshi: Moshi,
     dispatchers: CoroutineDispatchers,
 ): SideEffectViewModel<
-        Context,
+        FragmentActivity,
         FeedAllSideEffect,
         FeedAllViewState
-        >(dispatchers, FeedAllViewState.NoRecommendations), FeedFollowingViewModel, FeedRecommendationsViewModel
+        >(dispatchers, FeedAllViewState.Disabled), FeedFollowingViewModel, FeedRecommendationsViewModel
 {
 
     override val feedRecommendationsHolderViewStateFlow: MutableStateFlow<List<FeedRecommendation>> = MutableStateFlow(emptyList())
 
     init {
-        loadFeedRecommendations()
+        setRecommendationsVisibility()
+    }
+
+    private fun setRecommendationsVisibility() {
+        viewModelScope.launch(mainImmediate) {
+            feedRepository.recommendationsToggleStateFlow.collect { enabled ->
+                if (enabled) {
+                    loadFeedRecommendations()
+                } else {
+                    feedRecommendationsHolderViewStateFlow.value = listOf()
+                    updateViewState(FeedAllViewState.Disabled)
+                }
+            }
+        }
     }
 
     override fun loadFeedRecommendations() {
+        if (!feedRepository.recommendationsToggleStateFlow.value) {
+            return
+        }
+
         viewModelScope.launch(mainImmediate) {
+            feedRepository.getPodcastById(
+                FeedId(FeedRecommendation.RECOMMENDATION_PODCAST_ID)
+            )?.firstOrNull()?.let { podcast ->
+                if (podcast.isPlaying) {
+                    submitSideEffect(FeedAllSideEffect.RefreshWhilePlaying)
+                    return@launch
+                }
+            }
+
             updateViewState(FeedAllViewState.Loading)
 
             repositoryDashboard.getRecommendedFeeds().collect { feedRecommended ->
@@ -116,15 +148,41 @@ internal class FeedAllViewModel @Inject constructor(
                 return@launch
             }
 
-            var feedRecommendationParamsList: MutableList<String> = mutableListOf()
+            feedRepository.getPodcastById(
+                FeedId(FeedRecommendation.RECOMMENDATION_PODCAST_ID)
+            ).firstOrNull()?.let { podcast ->
 
-            for (r in recommendations) {
-                feedRecommendationParamsList.add(
-                    r.toJson(moshi)
+                pauseEpisodeIfNeeded(
+                    podcast,
+                    FeedId(feed.id)
+                )
+
+                dashboardNavigator.toCommonPlayerScreen(
+                    podcast.id,
+                    FeedId(feed.id),
+                    podcast.episodeDuration ?: 0
                 )
             }
+        }
+    }
 
-            dashboardNavigator.toCommonPlayerScreen(feedRecommendationParamsList, feed.id)
+    private fun pauseEpisodeIfNeeded(
+        podcast: Podcast,
+        episodeId: FeedId
+    ) {
+        viewModelScope.launch(mainImmediate) {
+            val currentEpisode = podcast.getCurrentEpisode()
+
+            if (currentEpisode.playing && currentEpisode.id != episodeId) {
+                podcast.didPausePlayingEpisode(currentEpisode)
+
+                mediaPlayerServiceController.submitAction(
+                    UserAction.ServiceAction.Pause(
+                        ChatId(ChatId.NULL_CHAT_ID.toLong()),
+                        currentEpisode.id.value
+                    )
+                )
+            }
         }
     }